refactor(lexer): split `carve` pipeline stage into multiple files
