+++ crates/oxc_lexer/src/pipeline/carve/mod.rs
@@ -1,0 +1,7 @@
+#[expect(clippy::module_inception)]
+mod carve;
+mod carve_jsx;
+mod common;
+
+pub use carve::carve;
+pub use carve_jsx::carve_jsx;
